refact imei-api
